split up storage.ts
