+++ src/application_logic/authEvent.ts
@@ -1,8 +1,8 @@
 import { changeUI } from "components/showApp";
-import { getProjects } from "./storage/storage";
 import projectStyles from "components/showProjects/showProjects.module.css";
 import { showProjectList } from "components/showProjects/showProjectList/showProjectList";
 import { createRoutes } from "./router";
+import { getProjects } from "./storage/getProjects";
 
 const getProjectListDiv = () =>
   document.querySelector(`.${projectStyles.projectList}`)!;
