+++ src/application_logic/storage/getInboxProject.ts
@@ -1,0 +1,32 @@
+import { auth } from "application_logic/auth";
+import { addDoc, query, where, getDocs } from "firebase/firestore";
+import { Project } from "types";
+import { createProjectData } from "./createItem";
+import { projectConverter } from "./firestoreConverter";
+import { projects } from "./useDb";
+
+// add inbox project for new users
+const createInboxProject = () => {
+  const projectData = createProjectData("Inbox", true);
+  return addDoc(projects, projectData).then(
+    (newInboxRef) => new Project(newInboxRef, projectData)
+  );
+};
+
+const getInboxProject = async () => {
+  const q = query(
+    projects,
+    where("ownerID", "==", auth.currentUser?.uid),
+    where("isInbox", "==", true)
+  ).withConverter(projectConverter);
+  const snapshot = await getDocs(q);
+
+  // for new users inbox project needs to be added first
+  if (!snapshot.size) return await createInboxProject();
+
+  return snapshot.docs[0].data();
+};
+
+const isInbox = (project: Project) => project.data.isInbox === true;
+
+export { getInboxProject, isInbox };
