+++ src/components/showProjects/showProject/showProject.ts
@@ -3,10 +3,10 @@ import { Project } from "types";
 import styles from "./showProject.module.css";
 import globalStyles from "style.module.css";
 import { router } from "application_logic/router";
-import { isInbox } from "application_logic/storage/storage";
 import { showIcon } from "components/helpers/buttons/showIcon";
 import { renameBtn } from "components/helpers/buttons/renameBtn";
 import { toggleMobileMenu } from "components/showHeader/mobileMenuBtn/mobileMenuBtn";
+import { isInbox } from "application_logic/storage/getInboxProject";
 
 const showToday = (parent: Element) => {
   const projectDiv = document.createElement("div");
